address multiple codacy and codebeat issues
528: Exception when opening PP Order issue / receipt 

Task-Url: https://github.com/metasfresh/metasfresh-webui-api/issues/528
