+++ src/main/java/de/metas/ui/web/handlingunits/process/WEBUI_M_HU_CreateMaterialReceipt.java
@@ -8,7 +8,6 @@ import org.adempiere.ad.dao.IQueryBL;
 import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.util.GuavaCollectors;
 import org.adempiere.util.Services;
-import org.adempiere.util.lang.MutableInt;
 import org.adempiere.util.lang.impl.TableRecordReference;
 import org.compiere.Adempiere;
 import org.compiere.model.IQuery;
@@ -21,15 +20,12 @@ import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_M_ReceiptSchedule;
 import de.metas.handlingunits.receiptschedule.IHUReceiptScheduleBL;
 import de.metas.process.IProcessPrecondition;
-import de.metas.process.IProcessPreconditionsContext;
-import de.metas.process.JavaProcess;
 import de.metas.process.Param;
 import de.metas.process.ProcessPreconditionsResolution;
 import de.metas.process.RunOutOfTrx;
 import de.metas.ui.web.WebRestApiApplication;
 import de.metas.ui.web.handlingunits.HUEditorRow;
 import de.metas.ui.web.handlingunits.HUEditorView;
-import de.metas.ui.web.process.ViewAsPreconditionsContext;
 import de.metas.ui.web.process.adprocess.ViewBasedProcessTemplate;
 import de.metas.ui.web.view.IViewsRepository;
 import de.metas.ui.web.view.ViewId;
@@ -57,56 +53,8 @@ import de.metas.ui.web.window.model.DocumentCollection;
  * #L%
  */
 @Profile(WebRestApiApplication.PROFILE_Webui)
-public class WEBUI_M_HU_CreateMaterialReceipt extends JavaProcess implements IProcessPrecondition
+public class WEBUI_M_HU_CreateMaterialReceipt extends WEBUI_M_HU_Receipt_Base implements IProcessPrecondition
 {
-	@Override
-	public ProcessPreconditionsResolution checkPreconditionsApplicable(final IProcessPreconditionsContext context)
-	{
-		final ViewAsPreconditionsContext viewContext = ViewAsPreconditionsContext.castOrNull(context);
-		if (viewContext == null)
-		{
-			return ProcessPreconditionsResolution.rejectWithInternalReason("webui view not available");
-		}
-
-		if (viewContext.isNoSelection())
-		{
-			return ProcessPreconditionsResolution.rejectBecauseNoSelection();
-		}
-
-		final MutableInt checkedDocumentsCount = new MutableInt(0);
-		final ProcessPreconditionsResolution firstRejection = viewContext.getView(HUEditorView.class)
-				.streamByIds(viewContext.getSelectedDocumentIds())
-				.filter(document -> document.isPureHU())
-				//
-				.peek(document -> checkedDocumentsCount.incrementAndGet()) // count checked documents
-				.map(document -> rejectResolutionOrNull(document)) // create reject resolution if any
-				.filter(resolution -> resolution != null) // filter out those which are not errors
-				.findFirst()
-				.orElse(null);
-		if (firstRejection != null)
-		{
-			// found a record which is not eligible => don't run the process
-			return firstRejection;
-		}
-		if (checkedDocumentsCount.getValue() <= 0)
-		{
-			return ProcessPreconditionsResolution.rejectWithInternalReason("no eligible rows");
-		}
-
-		// Safe to run the process
-		return ProcessPreconditionsResolution.accept();
-	}
-
-	private static final ProcessPreconditionsResolution rejectResolutionOrNull(final HUEditorRow document)
-	{
-		if (!document.isHUStatusPlanning())
-		{
-			return ProcessPreconditionsResolution.reject("Only planning HUs can be received"); // TODO: trl
-		}
-
-		return null;
-	}
-
 	@Autowired
 	private IViewsRepository viewsRepo;
 	@Autowired
@@ -120,8 +68,20 @@ public class WEBUI_M_HU_CreateMaterialReceipt extends JavaProcess implements IPr
 
 	public WEBUI_M_HU_CreateMaterialReceipt()
 	{
-		super();
 		Adempiere.autowire(this);
+	}
+
+	/**
+	 * Only allows rows whose HU is in the "planning" status.
+	 */
+	@Override
+	final ProcessPreconditionsResolution rejectResolutionOrNull(final HUEditorRow document)
+	{
+		if (!document.isHUStatusPlanning())
+		{
+			return ProcessPreconditionsResolution.reject("Only planning HUs can be received"); // TODO: trl
+		}
+		return null;
 	}
 
 	@Override
