+++ src/main/java/de/metas/ui/web/handlingunits/process/WEBUI_M_HU_ReverseReceipt.java
@@ -8,7 +8,6 @@ import org.adempiere.ad.dao.IQueryFilter;
 import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.util.GuavaCollectors;
 import org.adempiere.util.Services;
-import org.adempiere.util.lang.MutableInt;
 import org.adempiere.util.lang.impl.TableRecordReference;
 import org.compiere.Adempiere;
 import org.compiere.model.I_M_InOut;
@@ -19,15 +18,12 @@ import de.metas.handlingunits.inout.ReceiptCorrectHUsProcessor;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_M_ReceiptSchedule;
 import de.metas.process.IProcessPrecondition;
-import de.metas.process.IProcessPreconditionsContext;
-import de.metas.process.JavaProcess;
 import de.metas.process.Param;
 import de.metas.process.ProcessPreconditionsResolution;
 import de.metas.process.RunOutOfTrx;
 import de.metas.ui.web.WebRestApiApplication;
 import de.metas.ui.web.handlingunits.HUEditorRow;
 import de.metas.ui.web.handlingunits.HUEditorView;
-import de.metas.ui.web.process.ViewAsPreconditionsContext;
 import de.metas.ui.web.process.adprocess.ViewBasedProcessTemplate;
 import de.metas.ui.web.view.IViewsRepository;
 import de.metas.ui.web.view.ViewId;
@@ -62,61 +58,8 @@ import de.metas.ui.web.window.model.DocumentCollection;
  *
  */
 @Profile(value = WebRestApiApplication.PROFILE_Webui)
-public class WEBUI_M_HU_ReverseReceipt extends JavaProcess implements IProcessPrecondition
+public class WEBUI_M_HU_ReverseReceipt extends WEBUI_M_HU_Receipt_Base implements IProcessPrecondition
 {
-	@Override
-	public ProcessPreconditionsResolution checkPreconditionsApplicable(final IProcessPreconditionsContext context)
-	{
-		final ViewAsPreconditionsContext viewContext = ViewAsPreconditionsContext.castOrNull(context);
-		if (viewContext == null)
-		{
-			return ProcessPreconditionsResolution.rejectWithInternalReason("webui view not available");
-		}
-
-		final boolean isHUView = viewContext.getView() instanceof HUEditorView;
-		if (!isHUView)
-		{
-			return ProcessPreconditionsResolution.rejectWithInternalReason("The current view is not an HUEditorView; view=" + viewContext.getView() + ";");
-		}
-
-		if (viewContext.isNoSelection())
-		{
-			return ProcessPreconditionsResolution.rejectBecauseNoSelection();
-		}
-
-		final MutableInt checkedDocumentsCount = new MutableInt(0);
-		final ProcessPreconditionsResolution firstRejection = viewContext.getView(HUEditorView.class)
-				.streamByIds(viewContext.getSelectedDocumentIds())
-				.filter(document -> document.isPureHU())
-				//
-				.peek(document -> checkedDocumentsCount.incrementAndGet()) // count checked documents
-				.map(document -> rejectResolutionOrNull(document)) // create reject resolution if any
-				.filter(resolution -> resolution != null) // filter out those which are not errors
-				.findFirst()
-				.orElse(null);
-		if (firstRejection != null)
-		{
-			// found a record which is not eligible => don't run the process
-			return firstRejection;
-		}
-		if (checkedDocumentsCount.getValue() <= 0)
-		{
-			return ProcessPreconditionsResolution.rejectWithInternalReason("no eligible rows");
-		}
-
-		return ProcessPreconditionsResolution.accept();
-	}
-
-	private static final ProcessPreconditionsResolution rejectResolutionOrNull(final HUEditorRow document)
-	{
-		if (!document.isHUStatusActive())
-		{
-			return ProcessPreconditionsResolution.reject("Only active HUs can be reversed"); // TODO: trl
-		}
-
-		return null;
-	}
-
 	@Autowired
 	private IViewsRepository viewsRepo;
 	@Autowired
@@ -129,10 +72,21 @@ public class WEBUI_M_HU_ReverseReceipt extends JavaProcess implements IProcessPr
 
 	public WEBUI_M_HU_ReverseReceipt()
 	{
-		super();
 		Adempiere.autowire(this);
 	}
 
+	/**
+	 * Only allows rows whose HUs are in the "active" status. 
+	 */
+	final ProcessPreconditionsResolution rejectResolutionOrNull(final HUEditorRow document)
+	{
+		if (!document.isHUStatusActive())
+		{
+			return ProcessPreconditionsResolution.reject("Only active HUs can be reversed"); // TODO: trl
+		}
+		return null;
+	}
+	
 	@Override
 	@RunOutOfTrx
 	protected String doIt() throws Exception
